Revert "[Deferred Shading] Implemented overall Deferred Shading Technique"
