+++ Shader/SkyBoxFragment.shader
@@ -1,4 +1,4 @@
-#version 460 core
+#version 330 core
 
 out vec4 FragColor;
