+++ Shader/SkyBoxVertex.shader
@@ -1,4 +1,4 @@
-#version 460 core
+#version 330 core
 
 layout(location = 0) in vec3 aPos;
