--- Shader/lightBoxFragment.shader
@@ -1,9 +1,0 @@
-#version 460 core
-layout (location = 0) out vec4 FragColor;
-
-uniform vec3 lightColor;
-
-void main()
-{           
-    FragColor = vec4(lightColor, 1.0);
-}
