+++ Shader/objectFragment.shader
@@ -1,4 +1,4 @@
-#version 460 core
+#version 330 core
 
 #define N_LIGHTS 3
 
@@ -55,7 +55,6 @@ void main()
 	if (u_IsTexture)
 	{
 		FragColor = vec4(finalColor, 1.0f) * texture(u_Texture, TextureCoord);
-		//FragColor = vec4(abs(Normal), 1.0f);
 	}
 	else
 	{
