+++ src/PointMass.h
@@ -1,16 +1,6 @@
 #pragma once
 
-#include <utility>
 #include <glm/glm.hpp>
-
-enum NeighborType
-{
-	STRUCTURAL_X,
-	STRUCTURAL_Y,
-	SHEAR,
-	FLEXION_X,
-	FLEXION_Y
-};
 
 class PointMass
 {
@@ -21,9 +11,6 @@ class PointMass
 		glm::vec3 velocity = glm::vec3(0.0f, 0.0f, 0.0f);
 		glm::vec3 acceleration = glm::vec3(0.0f, 0.0f, 0.0f);
 		glm::vec3 force = glm::vec3(0.0f, 0.0f, 0.0f);
-
-		std::pair<PointMass*, NeighborType> neighborList[12];
-		size_t neighborSize = 0;
 
 		PointMass(const glm::vec3 &position);
